refactor: abstract passwordFinder functionality
All index#.js files were renamed to passwordFinder#.js. New index.js will be the main entry point for the module.The passwordFinder files encapsulate functionality for the various implementations of the day 04 part 1 solution.

Abstract the regex in conditionCheck from the main module, and create isConsecutiveDuplicate utility.

Rename the num parameter of conditionCheck to numString throught  the function signature as this a more accurate description of the argument type.